pyproject.toml
[project]
name = "multi-agent-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.1.0",
    "langchain-core>=1.1.0",
    "langchain-openai>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-text-splitters>=1.0.0",
    "faiss-cpu>=1.7.4",
    "python-dotenv>=1.2.1",
    "unstructured>=0.10.0",
    "markdown>=3.4.0",
]
